refactor(runtime): split bigint.rs into a bigint/ module directory
bigint.rs reached 2201 lines and tripped the 2000-line check_file_size.sh
lint gate. Split it into a module directory mirroring object/ and gc/:

  mod.rs      core: constants, error throwers, BigIntHeader, allocation,
              clean_bigint_ptr, and the shared limb primitives
  convert.rs  constructors, coercion, string parsing, outward conversions
  arith.rs    neg/not/add/sub/mul/div/mod/pow
  bitwise.rs  shifts, and/or/xor, asIntN/asUintN
  compare.rs  cmp/eq, string_to_bigint, bigint_cmp_f64
  tests.rs    unit tests (moved verbatim)

Pure refactor: no behavior change. All 32 #[no_mangle] exports and the
crate::bigint::* Rust surface are preserved via `pub use`. Re-points the
addr-class ratchet baseline for the moved handle-floor sites
(bigint.rs | 4 -> bigint/convert.rs | 2 + bigint/mod.rs | 2).
