+++ scripts/addr_class_ratchet_baseline.txt
@@ -33,7 +33,8 @@ handle-floor | crates/perry-runtime/src/array/iter_object.rs | 1
 handle-floor | crates/perry-runtime/src/array/iterator.rs | 2
 handle-floor | crates/perry-runtime/src/array/push_pop.rs | 1
 handle-floor | crates/perry-runtime/src/async_hooks.rs | 1
-handle-floor | crates/perry-runtime/src/bigint.rs | 4
+handle-floor | crates/perry-runtime/src/bigint/convert.rs | 2
+handle-floor | crates/perry-runtime/src/bigint/mod.rs | 2
 handle-floor | crates/perry-runtime/src/box.rs | 2
 handle-floor | crates/perry-runtime/src/buffer/access.rs | 1
 handle-floor | crates/perry-runtime/src/buffer/cmp.rs | 2
